// packages/cantonjs-splice-validator/src/generated/validatorInternal.types.ts
export interface operations {
    isReady: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description service_unavailable */
            503: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
        };
    };
    isLive: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description service_unavailable */
            503: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
        };
    };
    getValidatorUserInfo: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["GetValidatorUserInfoResponse"];
                };
            };
        };
    };
    register: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: {
            content: {
                "application/json": components["schemas"]["RegistrationRequest"];
            };
        };
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["RegistrationResponse"];
                };
            };
        };
    };
    listUsers: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ListUsersResponse"];
                };
            };
        };
    };
    onboardUser: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": components["schemas"]["OnboardUserRequest"];
            };
        };
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["OnboardUserResponse"];
                };
            };
        };
    };
    offboardUser: {
        parameters: {
            query: {
                username: string;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            404: components["responses"]["404"];
        };
    };
    dumpParticipantIdentities: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["NodeIdentitiesDump"];
                };
            };
        };
    };
    getDecentralizedSynchronizerConnectionConfig: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["GetDecentralizedSynchronizerConnectionConfigResponse"];
                };
            };
        };
    };
    getValidatorDomainDataSnapshot: {
        parameters: {
            query: {
                /**
                 * @description The timestamp as of which the dump (in particular, the ACS) is valid.
                 *
                 *     Must in the ISO-8601 format in UTC timezone, e.g.,
                 *     `yyyy-MM-dd'T'HH:mm:ss.SSS'Z'`.
                 */
                timestamp: string;
                /** @description The current migration id. */
                migration_id?: number;
                /**
                 * @description If true, do not check whether the provided timestamp is clean.
                 *     Not recommended for production,
                 *     see the `ExportAcs` endpoint of the `ParticipantRepairService` participant gRPC API.
                 */
                force?: boolean;
            };
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["GetValidatorDomainDataSnapshotResponse"];
                };
            };
            500: components["responses"]["500"];
        };
    };
    lookupTransferPreapprovalByParty: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                "receiver-party": string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["LookupTransferPreapprovalByPartyResponse"];
                };
            };
            404: components["responses"]["404"];
        };
    };
    cancelTransferPreapprovalByParty: {
        parameters: {
            query?: never;
            header?: never;
            path: {
                "receiver-party": string;
            };
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            404: components["responses"]["404"];
        };
    };
    listTransferPreapprovals: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["ListTransferPreapprovalsResponse"];
                };
            };
        };
    };
    prepareTransferPreapprovalSend: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": components["schemas"]["PrepareTransferPreapprovalSendRequest"];
            };
        };
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["PrepareTransferPreapprovalSendResponse"];
                };
            };
            400: components["responses"]["400"];
            404: components["responses"]["404"];
            501: components["responses"]["501"];
        };
    };
    submitTransferPreapprovalSend: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": components["schemas"]["SubmitTransferPreapprovalSendRequest"];
            };
        };
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["SubmitTransferPreapprovalSendResponse"];
                };
            };
            400: components["responses"]["400"];
            404: components["responses"]["404"];
            501: components["responses"]["501"];
        };
    };
    generateExternalPartyTopology: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": components["schemas"]["GenerateExternalPartyTopologyRequest"];
            };
        };
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["GenerateExternalPartyTopologyResponse"];
                };
            };
            501: components["responses"]["501"];
        };
    };
    submitExternalPartyTopology: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody: {
            content: {
                "application/json": components["schemas"]["SubmitExternalPartyTopologyRequest"];
            };
        };
        responses: {
            /** @description ok */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["SubmitExternalPartyTopologyResponse"];
                };
            };
            501: components["responses"]["501"];
        };
    };
// ...
}